+++ src/main/java/com/olyno/gami/models/Game.java
@@ -7,9 +7,11 @@ import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.LinkedList;
+import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -21,58 +23,101 @@ import com.olyno.gami.enums.GameMessageType;
 import com.olyno.gami.enums.GameState;
 import com.olyno.gami.listeners.GameListener;
 
+import org.yaml.snakeyaml.DumperOptions;
 import org.yaml.snakeyaml.Yaml;
-import org.yaml.snakeyaml.constructor.Constructor;
 
-public class Game extends GameManager {
+public class Game extends GameManager implements Cloneable {
 
-	private HashMap<String, Team> teams;
+	private ArrayList<Team> teams;
 	private GameMessageAs timerMessageAs;
 	private ArrayList<Game> sessions;
+	private boolean isSession;
 
+	private int id;
 	private GameState state;
 	private Integer timer;
 	private Team winner;
 	private Object world;
 
+	/**
+	 * Create a new Game.
+	 * Add the game in the list only if it already doesn't exist.
+	 * 
+	 * @param name The name of the game
+	 */
 	public Game(String name) {
 		super(name);
+		this.id = Gami.getGames().size();
 		this.state = GameState.WAITING;
 		this.timer = 15;
 		this.timerMessageAs = GameMessageAs.TITLE;
-		this.teams = new HashMap<>();
+		this.teams = new ArrayList<>();
 		this.sessions = new ArrayList<>();
-		this.getMessages(GameMessageType.TIMER)
-			.add(new GameTimerMessage(20, GameMessageTarget.GLOBAL, "Game starts in ${time} seconds"));
-		this.getMessages(GameMessageType.TIMER)
-			.add(new GameTimerMessage(15, GameMessageTarget.GLOBAL, "Game starts in ${time} seconds"));
+		this.isSession = false;
+		this.addMessage(GameMessageType.TIMER,
+			new GameTimerMessage(20, GameMessageTarget.GLOBAL, "Game starts in ${time} seconds"));
+		this.addMessage(GameMessageType.TIMER,
+			new GameTimerMessage(15, GameMessageTarget.GLOBAL, "Game starts in ${time} seconds"));
 		for (int time = 1; time < 11; time++) {
-			this.getMessages(GameMessageType.TIMER)
-				.add(new GameTimerMessage(time, GameMessageTarget.GLOBAL, "Game starts in ${time} seconds"));
+			this.addMessage(GameMessageType.TIMER,
+				new GameTimerMessage(time, GameMessageTarget.GLOBAL, "Game starts in ${time} seconds")
+			);
 		}
-		this.getMessages(GameMessageType.JOIN)
-			.add(new GameMessage(GameMessageTarget.GLOBAL, "${player} joined the game!"));
-		this.getMessages(GameMessageType.JOIN)
-			.add(new GameMessage(GameMessageTarget.PLAYER, "You joined the game ${game}"));
-		this.getMessages(GameMessageType.LEAVE)
-			.add(new GameMessage(GameMessageTarget.GLOBAL, "${player} left the game!"));
-		this.getMessages(GameMessageType.LEAVE)
-			.add(new GameMessage(GameMessageTarget.PLAYER, "You left the game ${game}"));
-		this.getMessages(GameMessageType.END).add(new GameMessage(GameMessageTarget.GLOBAL,
-			"The ${game} game is finished! The winner is the ${winner} team!"));
-		if (!Gami.getGames().containsKey(name)) {
-			Gami.getGames().put(name, this);
+		this.addMessage(GameMessageType.JOIN,
+			new GameMessage(GameMessageTarget.GLOBAL, "${player} joined the game!"));
+		this.addMessage(GameMessageType.JOIN,
+			new GameMessage(GameMessageTarget.PLAYER, "You joined the game ${game}"));
+		this.addMessage(GameMessageType.LEAVE,
+			new GameMessage(GameMessageTarget.GLOBAL, "${player} left the game!"));
+		this.addMessage(GameMessageType.LEAVE,
+			new GameMessage(GameMessageTarget.PLAYER, "You left the game ${game}"));
+		this.addMessage(GameMessageType.END,
+			new GameMessage(GameMessageTarget.GLOBAL, "The ${game} game is finished! The winner is the ${winner} team!"));
+		if (!Gami.getGameByName(this.name).isPresent()) {
+			Gami.getGames().add(this);
 			for (GameListener listener : Gami.getGameListeners()) {
 				listener.onGameCreated(this);
 			}
 		}
 	}
 
+	@Override
+	public boolean equals(Object o) {
+		if (o == null) return false;
+		if (o == this) return true;
+		if (!(o instanceof Game)) return false;
+		Game game = (Game) o;
+		return game.getName() == this.getName();
+	}
+
+	@Override
+	public int hashCode() {
+		return name.hashCode();
+	}
+
+	/**
+	 * Update the list with all games.
+	 * If regreshAll is true, will update id of all games.
+	 * 
+	 * @param refreshAll Should we refresh all ids?
+	 */
+	private Void update(Boolean refreshAll) {
+		return CompletableFuture.runAsync(() -> {
+			if (refreshAll) {
+				for (int i = 0; i < Gami.getGames().size(); i++) {
+					Gami.getGames().get(i).id = i;
+				}
+			} else {
+				Gami.getGames().set(this.id, this);
+			}
+		}).join();
+	}
+
 	/**
 	 * Starts the game
 	 */
 	public void start() {
-		this.state = GameState.START;
+		this.setState(GameState.START);
 		for (GameListener listener : Gami.getGameListeners()) {
 			listener.onGameStarted(this);
 		}
@@ -83,6 +128,7 @@ public class Game extends GameManager {
 	 */
 	public void stop() {
 		this.state = GameState.ENDED;
+		this.update(true);
 		for (GameListener listener : Gami.getGameListeners()) {
 			listener.onGameStopped(this);
 		}
@@ -92,10 +138,21 @@ public class Game extends GameManager {
 	 * Delete the game
 	 */
 	public void delete() {
-		Gami.getGames().remove(this.name);
+		Gami.getGames().remove(this);
+		this.update(true);
 		for (GameListener listener : Gami.getGameListeners()) {
 			listener.onGameDeleted(this);
 		}
+	}
+
+	/**
+	 * Alias of Game#save(Path, FileFormat) with other types.
+	 * 
+	 * @param directory The directory where the file should be saved
+	 * @param format    The format of the file (yaml, json...)
+	 */
+	public void save(String directory, FileFormat format) {
+		this.save(Paths.get(directory), format);
 	}
 
 	/**
@@ -106,38 +163,42 @@ public class Game extends GameManager {
 	 * @param format    The format of the file (yaml, json...)
 	 */
 	public void save(Path directory, FileFormat format) {
-		if (Files.isDirectory(directory)) {
-			try {
-				Path gameFile = Paths.get(this.name + "." + format.name().toLowerCase());
-				String data;
-				switch (format) {
-					case YAML:
-						Yaml yaml = new Yaml(new Constructor(Game.class));
-						data = yaml.dump(this);
-						break;
-					case JSON:
-						Gson gson = new GsonBuilder().setPrettyPrinting().create();
-						data = gson.toJson(this);
-						break;
-					default:
-						throw new Error("This format doesn't exist: " + format);
-				}
-				Files.write(gameFile, Arrays.asList(data.split("\n")), StandardOpenOption.CREATE);
-				for (GameListener listener : Gami.getGameListeners()) {
-					listener.onGameSaved(this, directory, format);
-				}
-			} catch (IOException ex) {
-				ex.printStackTrace();
+		try {
+			Path gameFile = Paths.get(directory.toString(), this.name + "." + format.name().toLowerCase());
+			String data = "";
+			switch (format) {
+				case YAML:
+					DumperOptions options = new DumperOptions();
+					options.setIndent(2);
+					options.setPrettyFlow(true);
+					// Fix below - additional configuration
+					options.setDefaultFlowStyle(DumperOptions.FlowStyle.BLOCK);
+					Yaml yaml = new Yaml(options);
+					data = yaml.dump(this);
+					break;
+				case JSON:
+					Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
+					data = gson.toJson(this);
+					break;
+				default:
+					throw new Error("This format doesn't exist: " + format);
 			}
+			Files.createDirectories(directory);
+			Files.write(gameFile, Arrays.asList(data.split("\n")), StandardOpenOption.CREATE);
+			for (GameListener listener : Gami.getGameListeners()) {
+				listener.onGameSaved(this, directory, format);
+			}
+		} catch (IOException ex) {
+			ex.printStackTrace();
 		}
 	}
 
 	/**
-	 * Returns a hashmap of all existing teams of this game
+	 * Returns an ArrayList of all existing teams of this game
 	 * 
 	 * @return List of teams
 	 */
-	public HashMap<String, Team> getTeams() {
+	public ArrayList<Team> getTeams() {
 		return teams;
 	}
 
@@ -157,27 +218,51 @@ public class Game extends GameManager {
 	 */
 	public void setState(GameState state) {
 		this.state = state;
+		this.update(false);
+	}
+
+	/**
+     * Returns a list of teams dependening of a filter.
+     * 
+	 * @return A list of teams dependening of a filter
+	 */
+	public ArrayList<Team> getTeamsByFilter(Predicate<? super Team> filter) {
+		List<Team> teamsFound = teams.stream().filter(filter).collect(Collectors.toList()); 
+		return new ArrayList<Team>(teamsFound);
+	}
+
+	/**
+     * Returns the first team which match with the filter.
+     * 
+	 * @return The first team which match with the filter
+	 */
+	public Optional<Team> getTeamByFilter(Predicate<? super Team> filter) {
+		return teams.stream().filter(filter).findFirst();
 	}
 
 	/**
 	 * Returns a team from its name
 	 *
-	 * @param team The team which will be return
-	 * @return Team with the name in argument
+	 * @param team The name of the team which will be return
+	 * @return Team with the same name
 	 */
-	public Team getTeam(String team) {
-		return teams.getOrDefault(team, null);
+	public Optional<Team> getTeamByName(String teamName) {
+		return teams.stream().filter(team -> team.getName().equals(teamName)).findFirst();
 	}
 
 	/**
 	 * Add a team to the game
+	 * Add the team in the list only if it already doesn't exist.
 	 *
 	 * @param team The team which will be added
 	 */
-	public void addTeam(Team team) {
-		teams.put(team.getName(), team);
-		for (GameListener listener : Gami.getGameListeners()) {
-			listener.onTeamAdded(this, team);
+	public void addTeam(Team teamToAdd) {
+		if (!teams.contains(teamToAdd)) {
+			teams.add(teamToAdd);
+			this.update(false);
+			for (GameListener listener : Gami.getGameListeners()) {
+				listener.onTeamAdded(this, teamToAdd);
+			}
 		}
 	}
 
@@ -187,7 +272,8 @@ public class Game extends GameManager {
 	 * @param team The team which will be removed
 	 */
 	public void removeTeam(Team team) {
-		teams.remove(team.getName());
+		teams.remove(team);
+		this.update(false);
 		for (GameListener listener : Gami.getGameListeners()) {
 			listener.onTeamRemoved(this, team);
 		}
@@ -199,8 +285,8 @@ public class Game extends GameManager {
 	 * @param team The team that you need to check existence
 	 * @return If the team exists or not
 	 */
-	public Boolean teamExists(String team) {
-		return teams.containsKey(team);
+	public Boolean teamExists(String teamName) {
+		return teams.stream().anyMatch(team -> team.getName().equals(teamName));
 	}
 
 	/**
@@ -219,6 +305,7 @@ public class Game extends GameManager {
 	 */
 	public <T> void setWorld(T world) {
 		this.world = world;
+		this.update(false);
 	}
 
 	/**
@@ -237,6 +324,7 @@ public class Game extends GameManager {
 	 */
 	public void setTimer(Integer time) {
 		this.timer = time;
+		this.update(false);
 	}
 
 	/**
@@ -255,6 +343,7 @@ public class Game extends GameManager {
 	 */
 	public void setTimerMessageAs(GameMessageAs as) {
 		this.timerMessageAs = as;
+		this.update(false);
 	}
 
 	/**
@@ -273,6 +362,7 @@ public class Game extends GameManager {
 	 */
 	public void setWinner(Team winner) {
 		this.winner = winner;
+		this.update(false);
 	}
 
 	/**
@@ -285,7 +375,8 @@ public class Game extends GameManager {
 	}
 
 	/**
-	 * Returns a specific game session from its id. Can be empty if the id is wrong or out of the range.
+	 * Returns a specific game session from its id. Can be empty if the id is wrong
+	 * or out of the range.
 	 * 
 	 * @return The session with its id
 	 */
@@ -298,14 +389,17 @@ public class Game extends GameManager {
 
 	/**
 	 * Create a new session for this game It's simply a copy of the game to host the
-	 * same game multiple times. Can be an empty optional if the session creation fail.
+	 * same game multiple times. Can be an empty optional if the session creation
+	 * fail.
 	 *
 	 * @return The created session
 	 */
 	public Optional<Game> createSession() {
 		try {
 			Game clonedGame = (Game) this.clone();
+			clonedGame.isSession = true;
 			sessions.add(clonedGame);
+			this.update(false);
 			for (GameListener listener : Gami.getGameListeners()) {
 				listener.onSessionCreated(clonedGame);
 			}
@@ -319,23 +413,34 @@ public class Game extends GameManager {
 	/**
 	 * Delete an exisiting session for this game
 	 *
+	 * @param id The id of the session to delete
 	 * @return The deleted session
 	 */
 	public Game deleteSession(Integer id) {
-		Game deletedGame = sessions.get(id);
+		Game deletedSession = sessions.get(id);
 		sessions.remove(id);
+		this.update(false);
 		for (GameListener listener : Gami.getGameListeners()) {
-			listener.onSessionDeleted(deletedGame);
+			listener.onSessionDeleted(deletedSession);
 		}
-		return deletedGame;
+		return deletedSession;
+	}
+
+	/**
+	 * Returns a boolean depending if the game is a session or not.
+	 *
+	 * @return A boolean depending if the game is a session or not.
+	 */
+	public Boolean isSession() {
+		return this.isSession;
 	}
 
 	@Override
-	@SuppressWarnings("unchecked")
 	public <T> void addPlayer(T player) {
 		if (!players.contains(player)) {
 			if (players.size() < maxPlayer) {
-				((LinkedList<T>) players).add(player);
+				players.add(player);
+				this.update(false);
 				for (GameListener listener : Gami.getGameListeners()) {
 					if (players.size() == minPlayer) {
 						listener.onGameCanStart(this);
@@ -355,6 +460,7 @@ public class Game extends GameManager {
 	public <T> void removePlayer(T player) {
 		if (players.contains(player)) {
 			players.remove(player);
+			this.update(false);
 			for (GameListener listener : Gami.getGameListeners()) {
 				listener.onPlayerLeave(this, player);
 			}
@@ -362,10 +468,10 @@ public class Game extends GameManager {
 	}
 
 	@Override
-	@SuppressWarnings("unchecked")
 	public <T> void addSpectator(T player) {
 		if (!spectators.contains(player)) {
-			((LinkedList<T>) this.spectators).add(player);
+			spectators.add(player);
+			this.update(false);
 			for (GameListener listener : Gami.getGameListeners()) {
 				listener.onSpectatorJoin(this, player);
 			}
@@ -376,6 +482,7 @@ public class Game extends GameManager {
 	public <T> void removeSpectator(T player) {
 		if (spectators.contains(player)) {
 			spectators.remove(player);
+			this.update(false);
 			for (GameListener listener : Gami.getGameListeners()) {
 				listener.onSpectatorLeave(this, player);
 			}