+++ src/main/java/com/olyno/gami/models/Team.java
@@ -1,7 +1,7 @@
 package com.olyno.gami.models;
 
+import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.LinkedList;
 import java.util.Optional;
 
 import com.olyno.gami.Gami;
@@ -11,34 +11,55 @@ import com.olyno.gami.listeners.TeamListener;
 
 public class Team extends GameManager {
 
-	private LinkedList<Point> points;
+	private ArrayList<Point> points;
 	private Integer goal;
 
-    /**
-     * Create a new team
+	/**
+	 * Create a new team
 	 * 
-     * @param name The name of the team
-     */
+	 * @param name The name of the team
+	 */
 	public Team(String name) {
 		super(name);
-        this.points = new LinkedList<Point>(Arrays.asList( new Point(0) ));
+		this.points = new ArrayList<Point>(Arrays.asList(new Point(0)));
 		this.goal = 5;
-		this.getMessages(GameMessageType.JOIN).add(new GameMessage(GameMessageTarget.GLOBAL, "${player} joined the ${team} team!"));
-		this.getMessages(GameMessageType.JOIN).add(new GameMessage(GameMessageTarget.PLAYER, "You joined the ${team} team!"));
-		this.getMessages(GameMessageType.LEAVE).add(new GameMessage(GameMessageTarget.GLOBAL, "${player} left the ${team} team!"));
-		this.getMessages(GameMessageType.LEAVE).add(new GameMessage(GameMessageTarget.PLAYER, "You left the ${team} team!"));
-		this.getMessages(GameMessageType.WIN_POINT).add(new GameMessage(GameMessageTarget.GLOBAL, "${player} scored a point for the ${team} team!"));
-		this.getMessages(GameMessageType.WIN_POINT).add(new GameMessage(GameMessageTarget.PLAYER, "You scored a point for your team!"));
-		this.getMessages(GameMessageType.LOSE_POINT).add(new GameMessage(GameMessageTarget.GLOBAL, "${player} lost a point for the ${team} team"));
-		this.getMessages(GameMessageType.LOSE_POINT).add(new GameMessage(GameMessageTarget.PLAYER, "The opponent team won a point!"));
+		this.addMessage(GameMessageType.JOIN,
+			new GameMessage(GameMessageTarget.GLOBAL, "${player} joined the ${team} team!"));
+		this.addMessage(GameMessageType.JOIN,
+			new GameMessage(GameMessageTarget.PLAYER, "You joined the ${team} team!"));
+		this.addMessage(GameMessageType.LEAVE,
+			new GameMessage(GameMessageTarget.GLOBAL, "${player} left the ${team} team!"));
+		this.addMessage(GameMessageType.LEAVE,
+			new GameMessage(GameMessageTarget.PLAYER, "You left the ${team} team!"));
+		this.addMessage(GameMessageType.WIN_POINT,
+			new GameMessage(GameMessageTarget.GLOBAL, "${player} scored a point for the ${team} team!"));
+		this.addMessage(GameMessageType.WIN_POINT,
+			new GameMessage(GameMessageTarget.PLAYER, "You scored a point for your team!"));
+		this.addMessage(GameMessageType.LOSE_POINT,
+			new GameMessage(GameMessageTarget.GLOBAL, "${player} lost a point for the ${team} team"));
+		this.addMessage(GameMessageType.LOSE_POINT,
+			new GameMessage(GameMessageTarget.PLAYER, "The opponent team won a point!"));
 		for (TeamListener listener : Gami.getTeamListeners()) {
 			listener.onTeamCreated(this);
 		}
-    }
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (o == null) return false;
+		if (o == this) return true;
+		if (!(o instanceof Team)) return false;
+		Team team = (Team) o;
+		return team.getName() == this.getName();
+	}
+
+	@Override
+	public int hashCode() {
+		return name.hashCode();
+	}
 
 	/**
-	 * Returns the goal of the team,
-	 * How many points the team must to have to win
+	 * Returns the goal of the team, How many points the team must to have to win
 	 * 
 	 * @return The goal of the team
 	 */
@@ -55,35 +76,32 @@ public class Team extends GameManager {
 		this.goal = goal;
 	}
 
-    /**
+	/**
 	 * Returns the list of points of a team
 	 * 
-     * @return The points of the team
-     */
-    public LinkedList<Point> getPoints() {
-        return points;
+	 * @return The points of the team
+	 */
+	public ArrayList<Point> getPoints() {
+		return points;
 	}
-	
+
 	/**
 	 * Returns the amount of points of a team
 	 * 
-     * @return The points of the team
-     */
-    public Integer getTotalPoints() {
-		return points
-			.stream()
-			.mapToInt(point -> point.getPoints())
-			.sum();
-    }
+	 * @return The points of the team
+	 */
+	public Integer getTotalPoints() {
+		return points.stream().mapToInt(point -> point.getPoints()).sum();
+	}
 
-    /**
-     * Set the points of the team
+	/**
+	 * Set the points of the team
 	 * 
-     * @param points The new points of the team
-     */
-    public void setPoints( Integer points ) {
-        this.points = new LinkedList<Point>(Arrays.asList( new Point(0) ));
-    }
+	 * @param points The new points of the team
+	 */
+	public void setPoints(Integer points) {
+		this.points = new ArrayList<Point>(Arrays.asList(new Point(0)));
+	}
 
 	/**
 	 * Set the points of the team
@@ -91,15 +109,15 @@ public class Team extends GameManager {
 	 * @param points The new points of the team
 	 */
 	public void setPoints(Point points) {
-		this.points = new LinkedList<Point>(Arrays.asList( points ));
+		this.points = new ArrayList<Point>(Arrays.asList(points));
 	}
 
-    /**
-     * Add points to the team
+	/**
+	 * Add points to the team
 	 * 
-     * @param points The points added
+	 * @param points The points added
 	 */
-	public void addPoints( Point points ) {
+	public void addPoints(Point points) {
 		this.points.add(points);
 		for (TeamListener listener : Gami.getTeamListeners()) {
 			listener.onPointWin(this, points);
@@ -113,12 +131,12 @@ public class Team extends GameManager {
 		}
 	}
 
-    /**
-     * Remove points from the team
+	/**
+	 * Remove points from the team
 	 * 
-     * @param points The point removed
+	 * @param points The point removed
 	 */
-	public void removePoints( Point points ) {
+	public void removePoints(Point points) {
 		this.points.remove(points);
 		for (TeamListener listener : Gami.getTeamListeners()) {
 			listener.onPointLose(this, points);
@@ -131,8 +149,8 @@ public class Team extends GameManager {
 	 * @return The game of the Team
 	 */
 	public Optional<Game> getGame() {
-		for (Game game : Gami.getGames().values()) {
-			if (game.getTeams().containsValue(this)) {
+		for (Game game : Gami.getGames()) {
+			if (game.getTeams().contains(this)) {
 				return Optional.of(game);
 			}
 		}
@@ -140,10 +158,9 @@ public class Team extends GameManager {
 	}
 
 	@Override
-	@SuppressWarnings("unchecked")
 	public <T> void addPlayer(T player) {
 		if (!players.contains(player)) {
-			((LinkedList<T>) players).add(player);
+			players.add(player);
 			for (TeamListener listener : Gami.getTeamListeners()) {
 				listener.onPlayerJoin(this, player);
 				if (minPlayer == players.size()) {
@@ -165,10 +182,9 @@ public class Team extends GameManager {
 	}
 
 	@Override
-	@SuppressWarnings("unchecked")
 	public <T> void addSpectator(T player) {
 		if (!spectators.contains(player)) {
-			((LinkedList<T>) spectators).add(player);
+			spectators.add(player);
 			for (TeamListener listener : Gami.getTeamListeners()) {
 				listener.onSpectatorJoin(this, player);
 			}
@@ -187,12 +203,12 @@ public class Team extends GameManager {
 
 	@Override
 	public void delete() {
-		if (getGame().isPresent()) {
-			getGame().get().getTeams().remove(this.getName());
+		getGame().ifPresent(game -> {
+			game.removeTeam(this);
 			for (TeamListener listener : Gami.getTeamListeners()) {
 				listener.onTeamDeleted(this);
 			}
-		}
+		});
 	}
 
 }