+++ src/test/java/com/olyno/gami/utils/PlayerTest.java
@@ -1,0 +1,29 @@
+package com.olyno.gami.utils;
+
+public class PlayerTest {
+
+    private String name;
+    private String displayName;
+
+    public PlayerTest(String name) {
+        this.name = name;
+        this.displayName = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    public void setDisplayName(String displayName) {
+        this.displayName = displayName;
+    }
+
+}